fix build error
Since the `master` branch of `date` is not on commit `77bd6b` anymore
but on `2709de`, the download fails with:

`module date:
Git commit for branch master is 2709de, but expected 77bd6b`

This change set updates the expected commit to the current master:
`2709deddd3c0deae74bc768b9d7c3b17dc4ba969`.

And since specifying the `master` branch and a commit hash as well
is fragile, this commit removes the reference to the `master` branch.